pyproject.toml
[tool.poetry]
name = "planner"
version = "0.1.0"
description = "手帳"
authors = ["nobituk <nobituk@ab.auone-net.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
reportlab = "^4.0.7"
pyyaml = "^6.0.1"
beautifulsoup4 = "^4.12.2"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pytest = "^7.4.3"
coverage = "^7.3.2"
grip = "^4.6.2"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
